Started work on the import for TV but ran into the fact that TV is structured different to movies, yet has some elements in common.
Looked at the API calls and started the tests for it, they still need work as they need the API for episodes adding.

Reviewed data and came up with a completely new model which will lead to re-write of much of the import too but will give a more solid logical model going forward, and will have the advantage of allowing search by actor or director later.